[954] MVP config-driven re-usable script for datasets validation
